♻️[#105] Refactor: SearchInput 액티비티 타입 삭제
