+++ packages/design-system/src/components/MainSearchInput/types/index.ts
@@ -1,0 +1,3 @@
+export interface SearchInputProps {
+  onClick: (keyword: string) => void;
+}
