+++ packages/design-system/src/pages/MainSearchInputDoc.tsx
@@ -1,6 +1,6 @@
 import Playground from '@/layouts/Playground';
 
-import MainSearchInput from '../components/MainSearchInput';
+import MainSearchInput from '../components/MainSearchInput/MainSearchInput';
 import DocTemplate, { DocCode } from '../layouts/DocTemplate';
 
 /* Playground에서 사용할 예시 코드 */
